requirements.txt
django-cors-headers==3.5.0
djangorestframework==3.12.1
mysqlclient>=1.4.6
django-filter==2.4.0
pycrypto==2.6.1
gunicorn==20.0.4
requests~=2.24.0
boto3==1.15.12
pika==1.1.0
pyheif==0.5.1
rsa==4.7.2
drf-yasg==1.21.5